pyproject.toml
[tool.poetry]
name = "dagflow"
version = "0.1.0"
description = "Ephemeral Airflow and DB instances for local development."
authors = ["Armaan Hussain <armaanPy>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
apache-airflow = "^2.9.3"
yfinance = "^0.2.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
